duplicate button component
