+++ app/layout/community/published-projects/detail/component.tsx
@@ -1,7 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 
 import { useRouter } from 'next/router';
-import { useToasts } from 'hooks/toast';
 
 import { format } from 'd3';
 
@@ -9,15 +8,11 @@ import { usePublishedProject } from 'hooks/projects';
 
 import Avatar from 'components/avatar';
 import Backlink from 'layout/statics/backlink';
-import Button from 'components/button';
-import Icon from 'components/icon';
+import DuplicateButton from 'layout/community/published-projects/duplicate-button';
 import Loading from 'components/loading';
 import PublishedProjectMap from 'layout/community/published-projects/detail/map';
 import Share from 'layout/community/published-projects/detail/share';
 import Wrapper from 'layout/wrapper';
-
-import CHECKED_SVG from 'svgs/ui/checked.svg?sprite';
-import DOWNLOAD_SVG from 'svgs/ui/download.svg?sprite';
 
 export interface CommunityProjectsDetailProps {
 
@@ -25,31 +20,16 @@ export interface CommunityProjectsDetailProps {
 
 export const CommunityProjectsDetail: React.FC<CommunityProjectsDetailProps> = () => {
   const { query } = useRouter();
-  const { addToast } = useToasts();
   const { pid } = query;
   const {
     data: publishedProject,
     isFetching: publishedProjectIsFetching,
     isFetched: publishedProjectIsFetched,
   } = usePublishedProject(pid);
-  const [isDuplicated, setIsDuplicated] = useState(false);
 
   const {
     description, name, planningAreaName, timesDuplicated, users, scenarios,
   } = publishedProject || {};
-
-  const handleDuplicated = () => {
-    setIsDuplicated(true);
-
-    addToast('success-upload-shapefile', (
-      <>
-        <h2 className="font-medium">Success!</h2>
-        <p className="text-sm">Project duplicated</p>
-      </>
-    ), {
-      level: 'success',
-    });
-  };
 
   return (
     <Wrapper>
@@ -78,18 +58,7 @@ export const CommunityProjectsDetail: React.FC<CommunityProjectsDetailProps> = (
               {description}
             </p>
             <div className="flex flex-row items-center mb-10">
-              <Button
-                className="px-6 group"
-                onClick={handleDuplicated}
-                size="s"
-                theme={isDuplicated ? 'white' : 'transparent-white'}
-              >
-                {isDuplicated ? 'Duplicated' : 'Duplicate'}
-                <Icon
-                  className={isDuplicated ? 'w-3.5 h-3.5 ml-2 text-black group-hover:text-white' : 'w-3.5 h-3.5 ml-2 text-white group-hover:text-black'}
-                  icon={isDuplicated ? CHECKED_SVG : DOWNLOAD_SVG}
-                />
-              </Button>
+              <DuplicateButton />
               {timesDuplicated && (
               <p className="ml-5 text-sm text-white">
                 Duplicated
