+++ app/layout/community/published-projects/duplicate-button/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
